add confirmation in the delete button

+++ app/dashboard/orders/page.tsx
@@ -133,8 +133,9 @@ const Page = (props: Props) => {
             <div className="my-4 flex gap-2">
                   <Button onClick={exportExcel}> Export Excel</Button>
                   <Button variant="outline" onClick={()=>setOrders(orders.map(o=>({...o,selected:false})))}> Unselect All</Button>
-                  <Button className="flex gap-2" onClick={()=>{
+                  <Button className="flex ml-auto gap-2" onClick={()=>{
 
+            if(!confirm("Are you sure you want to delete the selected orders?")) return
             orders
             .filter(o=>o.selected==true).forEach(o=>{
                 deleteDoc(doc(db,"orders",o.id)).then(()=>{
